fix(geth-types): address review findings — unsigned-underflow guards and dead checks
HIGH (consensus / correctness):
- EvmInstructions.Storage SSTORE refund: saturating subtract to prevent the ulong
  wrap that would silently grant the maximum post-cap refund
- StateSyncPivot.Diff: precedence bug returned the absolute number instead of the
  difference; reroute through SaturatingSub
- BlockTree.AcceptVisitor: bail on empty/inverted range instead of looping ~2^64
- BlockTreeSuggestPacer.WaitForQueue: guard the subtract so a head overtake
  doesn't pause the pacer indefinitely
- BinarySearchBlockNumber: bail at index==0 in the Down branch
- WitnessGeneratingHeaderFinder: count-driven loop; old i-- looped forever when
  _lowestRequestedHeader == 0

MEDIUM:
- SaturatingSub promoted to a shared UInt64Extensions; TransactionProcessor and
  EthereumGasPolicy now use it
- SszNumericChecks deleted; ulong fields no longer narrow to long
- ReceiptMigration.tx-index expiry: explicit-guard intent, no more (long?) wrap trick
- EraStore.GetEpochNumber: throw on blockNumber < FirstBlock
- ReceiptFixMigration: SaturatingSub Head?.Number - 2
- EraAdminRpcModule / EraE.AdminEraService: reject negative start/end before
  the (ulong) cast turns -1 into MaxValue
- E2StoreReader: reject negative starting_number from on-disk int64
- BlockParameter JSON Read: TryGetUInt64 path instead of throwing FormatException

Xdc (flcl42):
- SubnetPenaltyHandler.minBlockNumber / startRange: guard underflows before clamp
- EpochSwitchManager.estBlockNum: SaturatingSub before Math.Max
- Eip8037BlockGasInclusionCheck.CalculateBlockRegularGas: saturate the multi-step
  subtraction so a bookkeeping bug doesn't silently blow the block gas limit

Cleanups:
- NewPayloadHandler emoji: explicit if-tree (the < 0 branch was dead on ulong)
- DebugRpcModule.debug_getBlockRlp: remove dead blockNumber >= 0 check
- TraceStorePruner: early-return on block.Number <= _blockToKeep
- XdcRpcModule: dead BlockNumber < 0 checks → BlockNumber is null
- StateSyncPivot.TrySetNewBestHeader: drop the no-op Math.Max(x, 0)
- PowForwardHeaderProvider: drop the no-op outer Math.Max and the stale comment
- BlockDownloaderTests: drop the no-op Math.Max(0UL, …) noise
- SurgeGasPriceOracle.GetAverageGasUsagePerBlock: dead currentBlockNumber >= 0 and
  the unguarded i-- at i==0
- BlockhashStore.GetBlockHashFromState: drop dead requiredBlockNumber < 0
- ISurgeConfig.FeeHistoryBlockCount, L2GasUsageWindowSize, MaxGasLimitRatio:
  int → ulong (eliminates the casts at call sites)

+++ src/Nethermind/Nethermind.Taiko.Test/TxPoolContentListsTests.cs
@@ -131,9 +131,9 @@ public class TxPoolContentListsTests
         }
     }
 
-    [TestCase(10, 0)]
-    [TestCase(0, 1)]
-    public void MaxGasLimitRatio_FiltersHighGasLimitTransactions(int maxGasLimitRatio, int expectedTxCount)
+    [TestCase(10UL, 0)]
+    [TestCase(0UL, 1)]
+    public void MaxGasLimitRatio_FiltersHighGasLimitTransactions(ulong maxGasLimitRatio, int expectedTxCount)
     {
         Transaction tx = Build.A.Transaction
             .WithType(TxType.EIP1559)
@@ -178,11 +178,11 @@ public class TxPoolContentListsTests
         Assert.That(totalTxCount, Is.EqualTo(expectedTxCount));
     }
 
-    [TestCase(21_000UL, 120, 0, 2, TestName = "Batch is full by bytes")]
-    [TestCase(2_100_000UL, 100_000, 10, 1, TestName = "Surge filter rejects transaction")]
+    [TestCase(21_000UL, 120, 0UL, 2, TestName = "Batch is full by bytes")]
+    [TestCase(2_100_000UL, 100_000, 10UL, 1, TestName = "Surge filter rejects transaction")]
     public void GasUsed_IsNotInflated_WhenTxIsRejectedAfterExecution(
         ulong tx1GasLimit, int maxBytesPerTxList,
-        int surgeMaxGasLimitRatio, int expectedBatchCount)
+        ulong surgeMaxGasLimitRatio, int expectedBatchCount)
     {
         // tx1 is rejected after successful execution, tx2 is accepted
         Transaction tx1 = Build.A.Transaction
